+++ src/panza3/llm/__init__.py
@@ -1,5 +1,5 @@
 from .base import LLM, ChatHistoryType, MessageType
 from .local import PeftLLM, TransformersLLM
-from .ollama_llm import OllamaLLM
+from .ollama import OllamaLLM
 
 __all__ = ["LLM", "ChatHistoryType", "MessageType", "OllamaLLM", "TransformersLLM", "PeftLLM"]
